+++ lib/files/calculateTotalFilesSize.ts
